app/requirements.txt
Flask
joblib
pandas
kafka-python
psycopg2-binary
paramiko
python-dotenv

